still trying to finish header
Most of the html for header is done except for the mobile version at the end of the header (will work on this later after the other page is finished).

still working on the css part of the header, the utility bar seem to be done, currently working on the second bar, which is logo, search bar, ...
the third bar still look hideous

next commit will try to finish the css of the header so can move on the the body.